backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
python-dateutil==2.8.2
cloudinary==1.37.0
gunicorn==21.2.0
